tests/webapi-w3c-fileapi-tests/fileapi/FileReader_abort_when_readyState_DONE.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification, 
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list 
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice, 
  this list of conditions and the following disclaimer in the documentation 
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors 
  may be used to endorse or promote products derived from this work without 
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS" 
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT, 
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY 
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, 
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.  
  
Authors:
        Fan,Weiwei <weiwix.fan@intel.com>

-->

<html>
  <head>
    <title>File API Test: FileReader_abort_when_readyState_DONE</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2011/WD-FileAPI-20111020/#dfn-filereader" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if FileReader.abort is called when readyState is DONE that expecting FileReader.result is to be set null" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
    <input id="file" type="file" multiple size="80" />
    <h3>Please upload files</h3>
    <div id="log"></div>
    <script type="text/javascript">
        var t = async_test(document.title);
        setup({timeout : 3600000});

        var node = document.getElementById('file');
        node.onchange = t.step_func(function () {
            // obtain input element through DOM
            var file = node.files[0];
            getAsText(file);
        });

        function getAsText(readFile) {
            var reader = new FileReader();
            // Read file into memory as UTF-16
            reader.readAsText(readFile, "UTF-16");
            // Handle success event
            reader.onload = t.step_func(function loaded(evt) {
                if (reader.readyState == reader.DONE) {
                    reader.abort();
                    reader.onabort = t.step_func(function () {
                        // If readyState = EMPTY or if readyState = DONE set result to null when the abort() method is called
                        assert_true(reader.result === null, "The value of result is " + reader.result);
                        t.done();
                    });
                } else {
                    assert_unreached("read state error: " + reader.readyState);
                    t.done();
                }
            });
            // handle error event
            reader.onerror = t.step_func(function (e) {
                assert_unreached("readAsText() error: " + e.target.error.name);
                t.done();
            });
        }
    </script>
  </body>
</html>
